Refactoring
Modified the way in which effects are appiled by abilities, or by
anything else that can call static function for that matter.

Also examining a ways in which you can create library of effect
modifiers based on caster and target attributes.

I'm also going to make it less abstracted and generalized and more
focused.  More generic code can be found in RPGSystem and more focused
code in Game Folder.
I try to keep Game code as simple as possible, because it's main purpose
is to show a way of using framework, and providing examples of sample
solutions to problems you might encounter.

Also it should be easy enough to just take Game and modify it to your
needs.